+++ lib/build_eval/server/travis.rb
@@ -9,7 +9,7 @@ module BuildEval
 
       def build_result(name)
         repo_string = "#{@username}/#{name}"
-        has_failed = ::Travis::Repository.find(repo_string).last_build.failed?
+        has_failed = ::Travis::Repository.find(repo_string).recent_builds.first.failed?
         BuildEval::Result::BuildResult.create(
           build_name:  repo_string,
           status_name: has_failed ? "Failure" : "Success"
@@ -19,6 +19,7 @@ module BuildEval
       def to_s
         "Travis CI #{@username}"
       end
+
     end
 
   end