+++ spec/lib/build_eval/server/travis_pro_spec.rb
@@ -19,7 +19,7 @@ describe BuildEval::Server::TravisPro do
     subject { described_class.new(constructor_args) }
 
     it "establishes the GitHub auth token" do
-      expect(Travis::Pro).to receive(:github_auth).with(github_token)
+      expect(::Travis::Pro).to receive(:github_auth).with(github_token)
 
       subject
     end
@@ -28,45 +28,70 @@ describe BuildEval::Server::TravisPro do
 
   describe "#build_result" do
 
-    let(:build_name)        { "some_build_name" }
-    let(:travis_repository) { instance_double(Travis::Client::Repository) }
-    let(:travis_build)      { instance_double(Travis::Client::Build) }
-    let(:build_result)      { instance_double(BuildEval::Result::BuildResult) }
+    let(:build_name)             { "some_build_name" }
+    let(:travis_repository)      { instance_double(::Travis::Client::Repository) }
+    let(:recent_builds)          { (1..3).map { instance_double(::Travis::Client::Build) } }
+    let(:last_build)             { recent_builds.first }
+    let(:last_build_failed_flag) { false }
+    let(:build_result)           { instance_double(BuildEval::Result::BuildResult) }
 
     subject { travis_pro_server.build_result(build_name) }
 
     before(:example) do
-      allow(Travis::Pro::Repository).to receive(:find).and_return(travis_repository)
-      allow(travis_repository).to receive(:last_build).and_return(travis_build)
-      allow(travis_build).to receive(:failed?).and_return(false)
+      allow(::Travis::Pro::Repository).to receive(:find).and_return(travis_repository)
+      allow(travis_repository).to receive(:recent_builds).and_return(recent_builds)
+      allow(last_build).to receive(:failed?).and_return(last_build_failed_flag)
       allow(BuildEval::Result::BuildResult).to receive(:create).and_return(build_result)
     end
 
     it "retrieves the relevant Travis repository" do
-      expect(Travis::Pro::Repository).to receive(:find).with("#{username}/#{build_name}").and_return(travis_repository)
+      expect(::Travis::Pro::Repository).to receive(:find).with("#{username}/#{build_name}")
 
       subject
     end
 
-    it "retrieves the last build on the Travis repository" do
-      expect(travis_repository).to receive(:last_build).and_return(travis_build)
+    it "retrieves the recent builds from Travis" do
+      expect(travis_repository).to receive(:recent_builds)
 
       subject
     end
 
-    it "retrieves the build status from the build" do
-      expect(travis_build).to receive(:failed?).and_return(false)
+    it "retrieves the build status from the last build" do
+      expect(last_build).to receive(:failed?)
 
       subject
     end
 
-    it "creates a build result" do
-      expect(BuildEval::Result::BuildResult).to receive(:create).with(
-        build_name:  "#{username}/#{build_name}",
-        status_name: "Success"
-      ).and_return(build_result)
+    it "creates a build result whose build name is the path to the repository" do
+      expect(BuildEval::Result::BuildResult).to(
+        receive(:create).with(hash_including(build_name: "#{username}/#{build_name}"))
+      )
 
       subject
+    end
+
+    context "when the last build has passed" do
+
+      let(:last_build_failed_flag) { false }
+
+      it "creates a successful build result" do
+        expect(BuildEval::Result::BuildResult).to receive(:create).with(hash_including(status_name: "Success"))
+
+        subject
+      end
+
+    end
+
+    context "when the last build has failed" do
+
+      let(:last_build_failed_flag) { true }
+
+      it "creates a failed build result" do
+        expect(BuildEval::Result::BuildResult).to receive(:create).with(hash_including(status_name: "Failure"))
+
+        subject
+      end
+
     end
 
     it "returns the parsed build result" do